Implement bind-mounting /{dev,sys,proc} with an userspace fallback
While this can use either plain "mount --bind" if cap_sys_admin is
available (thanks to the "chkcap" tool), or fallback to the bindfs
FUSE program, it could work better because we are not really doing
a full chroot (due to fakechroot).

This needs some work before it's ready for prime time.