CI: Make build-wpebackend depend on upload-imagebuild
This is because the job uses aperezdc/ci-imagebuild:latest, and we would
rather build a fresh image using the newly built image. An issue remains:
building outside of the “master” skips building any image other than
ci-imagebuild because there's no reasonable way of saving the newly
built image to be passed along to the next job without uploading it
first.